requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
django-admin-sortable==2.1.18
django-admirarchy==1.0.0
django-allauth==0.40.0
django-appconf==1.0.3
django-avatar==4.1.0
django-bootstrap-carousel==0.2
django-braces==1.13.0
django-crispy-forms==1.8.0
django-debug-toolbar==2.0
django-etc==1.1.0
django-filer==1.6.0
django-fontawesome-5==1.0.16
django-haystack==2.8.1
django-hero-slider==1.3.1
django-hvad==1.8.0
django-jet==1.0.8
django-libs==2.0.2
django-localflavor==2.2
django-maintenance-mode==0.13.3
django-model-utils==3.2.0
django-mptt
django-nose==1.4.6
django-polymorphic==2.0.3
django-sitecats==1.0.0
django-sitecats-helpers==0.0.3
django-suit==0.2.28
django-tagging==0.4.6
easy-thumbnails==2.6
gunicorn==19.9.0
idna==2.8
midtransclient==1.1.0
mysqlclient==1.4.5
nose==1.3.7
oauthlib==3.1.0
Pillow==6.2.1
pysolr==3.8.1
python-decouple==3.1
python-stdnum==1.12
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.11.0
South==1.0.2
sqlparse==0.3.0
swapper==1.1.1
Unidecode==1.0.23
urllib3==1.25.6
whitenoise==4.1.4
